GCP/SFTP/cloud_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-pubsub
google-api-python-client>=2.0.0
